% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{fertilization}
\alias{fertilization}
\title{Height of 15 plant pairs collected by Darwin}
\format{
A data.frame with 2 columns and 15 observations.
}
\usage{
fertilization
}
\value{
a data.frame.
}
\description{
The data set contains heights of
self fertilized and cross fertilized plants as coded by Darwin.
}
\keyword{datasets}
