% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{is.RoBTT}
\alias{is.RoBTT}
\title{Reports whether x is a 'RoBTT' object}
\usage{
is.RoBTT(x)
}
\arguments{
\item{x}{an object to test}
}
\value{
\code{is.RoBTT} returns a boolean.
}
\description{
Reports whether x is a 'RoBTT' object
}
