% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.RoBTT}
\alias{plot.RoBTT}
\title{Plots a fitted 'RoBTT' object}
\usage{
\method{plot}{RoBTT}(
  x,
  parameter = "mu",
  conditional = FALSE,
  plot_type = "base",
  prior = FALSE,
  dots_prior = NULL,
  ...
)
}
\arguments{
\item{x}{a fitted 'RoBTT' object}

\item{parameter}{a parameter to be plotted. Defaults to
\code{"delta"} (for the effect size). The additional options
are \code{"rho"} (for the heterogeneity),
\code{"nu"} (for the degrees of freedom).}

\item{conditional}{whether conditional estimates should be
plotted. Defaults to \code{FALSE} which plots the model-averaged
estimates. Note that both \code{"weightfunction"} and
\code{"PET-PEESE"} are always ignoring the other type of
publication bias adjustment.}

\item{plot_type}{whether to use a base plot \code{"base"}
or ggplot2 \code{"ggplot"} for plotting. Defaults to
\code{"base"}.}

\item{prior}{whether prior distribution should be added to
figure. Defaults to \code{FALSE}.}

\item{dots_prior}{list of additional graphical arguments
to be passed to the plotting function of the prior
distribution. Supported arguments are \code{lwd},
\code{lty}, \code{col}, and \code{col.fill}, to adjust
the line thickness, line type, line color, and fill color
of the prior distribution respectively.}

\item{...}{list of additional graphical arguments
to be passed to the plotting function. Supported arguments
are \code{lwd}, \code{lty}, \code{col}, \code{col.fill},
\code{xlab}, \code{ylab}, \code{main}, \code{xlim}, \code{ylim}
to adjust the line thickness, line type, line color, fill color,
x-label, y-label, title, x-axis range, and y-axis range
respectively.}
}
\value{
\code{plot.RoBTT} returns either \code{NULL} if \code{plot_type = "base"}
or an object object of class 'ggplot2' if \code{plot_type = "ggplot2"}.
}
\description{
\code{plot.RoBTT} allows to visualize
different \code{"RoBTT"} object parameters in various
ways. See \code{type} for the different model types.
}
\examples{
\dontrun{
# using the example data from Anderson et al. 2010 and fitting the default model
# (note that the model can take a while to fit)
}


}
\seealso{
\code{\link[=RoBTT]{RoBTT()}}
}
