% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.RoBTT}
\alias{print.RoBTT}
\title{Prints a fitted RoBTT object}
\usage{
\method{print}{RoBTT}(x, ...)
}
\arguments{
\item{x}{a fitted RoBTT object.}

\item{...}{additional arguments.}
}
\value{
\code{print.RoBTT} invisibly returns the print statement.
}
\description{
Prints a fitted RoBTT object
}
\seealso{
\code{\link[=RoBTT]{RoBTT()}}
}
