% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.summary.RoBTT}
\alias{print.summary.RoBTT}
\title{Prints summary object for 'RoBTT' method}
\usage{
\method{print}{summary.RoBTT}(x, ...)
}
\arguments{
\item{x}{a summary of a 'RoBTT' object}

\item{...}{additional arguments}
}
\value{
\code{print.summary.RoBTT} invisibly returns the print statement.
}
\description{
Prints summary object for 'RoBTT' method
}
\seealso{
\code{\link[=RoBTT]{RoBTT()}}
}
