% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\docType{data}
\name{rho2logsdr}
\alias{rho2logsdr}
\title{rho to log standard deviation ratio transformations}
\format{
An object of class \code{list} of length 3.
}
\usage{
rho2logsdr
}
\value{
a list with the corresponding functions
}
\description{
A list containing the transformation function,
inverse transformation function, and the jacobian function.
}
\keyword{datasets}
