% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RoBTT-package.R
\docType{package}
\name{RoBTT-package}
\alias{RoBTT-package}
\alias{_PACKAGE}
\alias{RoBTT_package}
\alias{RoBTT.package}
\title{RoBTT: Robust Bayesian t-test}
\description{
RoBTT: Bayesian model-averaged t-test extends the Bayesian t-test by
incorporating inference about heterogeneity of variances and outliers.
}
\section{User guide}{
 See \insertCite{maier2022bayesian;textual}{RoBTT} for
details regarding the RoBTT methodology.

More details regarding customization of the Bayesian model-averaged t-test
are provided in
\href{../doc/Introduction_to_RoBTT.html}{\bold{Introduction to RoBTT}}
vignette. Please, use the "Issues" section in the GitHub repository to
ask any further questions.
}

\references{
\insertAllCited{}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://fbartos.github.io/RoBTT/}
  \item Report bugs at \url{https://github.com/FBartos/RoBTT/issues}
}

}
\author{
František Bartoš \email{f.bartos96@gmail.com}
}
\keyword{package}
