% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{RoBTT}
\alias{RoBTT}
\title{Estimate a Robust Bayesian T-Test}
\usage{
RoBTT(
  x1 = NULL,
  x2 = NULL,
  mean1 = NULL,
  mean2 = NULL,
  sd1 = NULL,
  sd2 = NULL,
  N1 = NULL,
  N2 = NULL,
  prior_delta = prior(distribution = "cauchy", parameters = list(location = 0, scale =
    sqrt(2)/2)),
  prior_rho = prior(distribution = "beta", parameters = list(alpha = 1, beta = 1)),
  prior_nu = prior(distribution = "exp", parameters = list(rate = 1)),
  prior_delta_null = prior(distribution = "spike", parameters = list(location = 0)),
  prior_rho_null = prior(distribution = "spike", parameters = list(location = 0.5)),
  prior_nu_null = NULL,
  likelihood = c("normal", if (!is.null(prior_nu)) "t"),
  chains = 4,
  iter = 10000,
  warmup = 5000,
  thin = 1,
  parallel = FALSE,
  control = set_control(),
  convergence_checks = set_convergence_checks(),
  save = "all",
  seed = NULL,
  silent = TRUE,
  ...
)
}
\arguments{
\item{x1}{vector of observations of the first group}

\item{x2}{vector of observations of the second group}

\item{mean1}{mean of the first group}

\item{mean2}{mean of the first group}

\item{sd1}{standard deviation of the first group}

\item{sd2}{standard deviation of the first group}

\item{N1}{sample size of the first group}

\item{N2}{sample size of the first group}

\item{prior_delta}{prior distributions for the effect size \code{delta} parameter
that will be treated as belonging to the alternative hypothesis. Defaults to \code{
prior(distribution = "Cauchy", parameters = list(location = 0, scale = sqrt(2)/2))}.}

\item{prior_rho}{prior distributions for the precision allocation \code{rho} parameter
that will be treated as belonging to the alternative hypothesis. Defaults to \code{
prior(distribution = "beta", parameters = list(alpha = 1, beta = 1))}.}

\item{prior_nu}{prior distribution for the degrees of freedom + 2 \code{nu}
parameter that will be treated as belonging to the alternative hypothesis.
Defaults to \code{prior(distribution = "exp", parameters = list(rate = 1))}.}

\item{prior_delta_null}{prior distribution for the \code{delta} parameter that
will be treated as belonging to the null hypothesis. Defaults to point distribution
with location at 0 (
\code{prior(distribution = "point", parameters = list(location = 0))}).}

\item{prior_rho_null}{prior distribution for the \code{rho} parameter that
will be treated as belonging to the null hypothesis. Defaults to point distribution
with location at 0.5 (
\code{prior(distribution = "point", parameters = list(location = 0.5))}).}

\item{prior_nu_null}{prior distribution for the \code{nu} parameter
that will be treated as belonging to the null hypothesis. Defaults to NULL (
\code{NULL}, i.e., normal distribution).}

\item{likelihood}{types of likelihood to be used. Defaults to \code{c("normal", "t")}
for a normal and t likelihoods.}

\item{chains}{a number of chains of the MCMC algorithm.}

\item{iter}{a number of sampling iterations of the MCMC algorithm.
Defaults to \code{10000}, with a minimum of \code{4000}.}

\item{warmup}{a number of warmup  iterations of the MCMC algorithm.
Defaults to \code{5000}.}

\item{thin}{a thinning of the chains of the MCMC algorithm. Defaults to
\code{1}.}

\item{parallel}{whether the individual models should be fitted in parallel.
Defaults to \code{FALSE}. The implementation is not completely stable
and might cause a connection error.}

\item{control}{allows to pass control settings with the
\code{\link[=set_control]{set_control()}} function. See \code{?set_control} for
options and default settings.}

\item{convergence_checks}{automatic convergence checks to assess the fitted
models, passed with \code{\link[=set_convergence_checks]{set_convergence_checks()}} function. See
\code{?set_convergence_checks} for options and default settings.}

\item{save}{whether all models posterior distributions should be kept
after obtaining a model-averaged result. Defaults to \code{"all"} which
does not remove anything. Set to \code{"min"} to significantly reduce
the size of final object, however, some model diagnostics and further
manipulation with the object will not be possible.}

\item{seed}{a seed to be set before model fitting, marginal likelihood
computation, and posterior mixing for reproducibility of results. Defaults
to \code{NULL} - no seed is set.}

\item{silent}{whether all print messages regarding the fitting process
should be suppressed. Defaults to \code{TRUE}. Note that \code{parallel = TRUE}
also suppresses all messages.}

\item{...}{additional arguments.}
}
\value{
\code{RoBTT} returns an object of \link[base]{class} \code{"RoBTT"}.
}
\description{
\code{RoBTT} is used to estimate a Robust Bayesian
T-Test. The input either requires the vector of observations for
each group, \code{x1, x2}, or the summary statistics (in case only
the \code{"normal"} likelihood is used).
}
\details{
See \insertCite{maier2022bayesian;textual}{RoBTT} for more details
regarding the methodology.

Generic \code{\link[=summary.RoBTT]{summary.RoBTT()}}, \code{\link[=print.RoBTT]{print.RoBTT()}}, and \code{\link[=plot.RoBTT]{plot.RoBTT()}} functions are
provided to facilitate manipulation with the ensemble.
}
\examples{
# using the example data from Darwin
data("fertilization", package = "RoBTT")
fit <- RoBTT(
  x1       = fertilization$Self,
  x2       = fertilization$Crossed,
  prior_delta = prior("cauchy", list(0, 1/sqrt(2))),
  prior_rho   = prior("beta",   list(3, 3)),
  likelihood  = "normal",
  seed        = 1,
  control     = set_control(adapt_delta = 0.95)
)

# summary can provide many details about the model
summary(fit)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link[=summary.RoBTT]{summary.RoBTT()}}, \code{\link[=prior]{prior()}}
}
