% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.RoBTT}
\alias{summary.RoBTT}
\title{Summarize fitted 'RoBTT' object}
\usage{
\method{summary}{RoBTT}(
  object,
  type = "ensemble",
  conditional = FALSE,
  group_estimates = FALSE,
  probs = c(0.025, 0.975),
  logBF = FALSE,
  BF01 = FALSE,
  short_name = FALSE,
  remove_spike_0 = FALSE,
  ...
)
}
\arguments{
\item{object}{a fitted 'RoBTT' object}

\item{type}{whether to show the overall 'RoBTT' results (\code{"ensemble"}),
an overview of the individual models (\code{"models"}), an overview of
the individual models MCMC diagnostics (\code{"diagnostics"}), or a detailed summary
of the individual models (\code{"individual"}). Can be abbreviated to first letters.}

\item{conditional}{show the conditional estimates (assuming that the
alternative is true). Defaults to \code{FALSE}. Only available for
\code{type == "conditional"}.}

\item{group_estimates}{show the model-averaged mean and standard deviation estimates for each group.}

\item{probs}{quantiles of the posterior samples to be displayed.
Defaults to \code{c(.025, .975)}}

\item{logBF}{show log of Bayes factors. Defaults to \code{FALSE}.}

\item{BF01}{show Bayes factors in support of the null hypotheses. Defaults to
\code{FALSE}.}

\item{short_name}{whether priors names should be shortened to the first
(couple) of letters. Defaults to \code{FALSE}.}

\item{remove_spike_0}{whether spike prior distributions with location at zero should
be omitted from the summary. Defaults to \code{FALSE}.}

\item{...}{additional arguments}
}
\value{
\code{summary.RoBTT} returns a list of tables of class 'BayesTools_table'.
}
\description{
\code{summary.RoBTT} creates summary tables for a
RoBTT object.
}
\seealso{
\code{\link[=RoBTT]{RoBTT()}}
}
